quadlet: Add config file
This adds support for config files in
 /usr/share/containers/quadlet.conf and/or
 /etc/containers/quadlet.conf, the later overriding the former.

Right now the only config is the "Dirs" location, which is initialized
to the default directories. This can be extended in the config file
with more directories, or overridden by first reseting with a "Dirs="
line.

The config format uses the systemd parser, so it is using the same
format as the quadlet files themselves, as well as other systemd
config files. Also, this makes the quadlet generator binary smaller
(having only one parser).

Signed-off-by: Alexander Larsson <alexl@redhat.com>