quadlet:Add support for signature verification
This adds two new options to the quadlet config: `RequireSignatures`,
which means all input files need valid signatures and `PublicKeyDirs`
that point to directories with Ed25519 public keys (in DER format) for
validation.

In addition, the config file can contain groups called `[Dir $path]`
with similar keys in them. The keys in the `Quadlet` group affect all
unit source directories, but the `Dir ...` ones only affect that
directory.

For source directories with RequireSignatures, each file `$file` must
be paired with a `$file.sig` file that contains the Ed25519 signature
of the file, and it must be validate agains one of the public keys
that are listed either globally or for this source directory.

The target usecase for this is having a read-only (immutable) system
(using e.g. ostree or dm-verity) with a transient /etc, but still want
to be able to install at run-time new system services using quadlet,
but only if the source of the quadlet is a trusted vendor. The later is
verified by validating a signature against a list of public keys that
are in the trusted system image.

For example, one could imagine an /etc/containers/quadlet.conf like:
```
[Quadlet]
Dir=/var/quadlets

[Dir /var/quadlets]
RequireSignatures=true
PublicKeyDirs=/etc/trusted-keys
```

This would then run any regular quadlets in /etc or /usr/share, but also
signed quadlets from /var/quadlets.

To test this, you can use openssl to generate a private and public key:

```
$ openssl genpkey -algorithm ed25519 -outform PEM -out testkey-secret.pem
$ openssl pkey -in testkey-secret.pem  -pubout -outform DER -out /etc/trusted-keys/testkey.pub
```

Then sign the quadlet:

```
$ openssl pkeyutl -sign -inkey testkey-secret.pem -rawin -in /var/quadlets/my.container  -out /var/quadlets/my.container.sig
```

Signed-off-by: Alexander Larsson <alexl@redhat.com>